backend/pyproject.toml
[project]
name = "research-stocks"
version = "0.1.0"
description = ""
authors = [
    {name = "Ricardo",email = "ricky_olalquiaga@hotmail.com"}
]
readme = "README.md"
requires-python = ">=3.12, <4.0"
dependencies = [
    "llama-index-llms-google-genai (>=0.7.3,<0.8.0)",
    "llama-index-tools-yahoo-finance (>=0.4.1,<0.5.0)",
    "llama-index-tools-tavily-research (>=0.4.2,<0.5.0)",
    "python-dotenv (>=1.2.1,<2.0.0)",
    "fastapi (>=0.122.0,<0.123.0)",
    "uvicorn (>=0.38.0,<0.39.0)",
    "mlflow (>=3.6.0,<4.0.0)",
    "pydantic-settings (>=2.12.0,<3.0.0)",
    "llama-index-embeddings-vertex (>=0.4.1,<0.5.0)",
    "apscheduler (>=3.11.1,<4.0.0)",
    "googlenews (>=1.6.15,<2.0.0)",
    "tradingview-ta (>=3.3.0,<4.0.0)"
]

[tool.poetry]
packages = [
    {include = "research_stocks", from = "src"},
    {include = "api", from = "src"},
    {include = "settings", from = "src"},
    {include = "utils", from = "src"},
    {include = "services", from = "src"}
    ]


[tool.poetry.group.dev.dependencies]
ipykernel = "^7.1.0"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
